Dev/E1000: Another trivial issue detected by parfait.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@82368 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Network/DevE1000.cpp
@@ -4534,11 +4534,11 @@ static int e1kFallbackAddToFrame(PPDMDEVINS pDevIns, PE1KSTATE pThis, E1KTXDESC 
     do
     {
         /* Calculate how many bytes we have left in this TCP segment */
-        uint32_t cb = u16MaxPktLen - pThis->u16TxPktLen;
+        uint16_t cb = u16MaxPktLen - pThis->u16TxPktLen;
         if (cb > pDesc->data.cmd.u20DTALEN)
         {
             /* This descriptor fits completely into current segment */
-            cb = pDesc->data.cmd.u20DTALEN;
+            cb = (uint16_t)pDesc->data.cmd.u20DTALEN; /* u20DTALEN at this point is guarantied to fit into 16 bits. */
             rc = e1kFallbackAddSegment(pDevIns, pThis, pDesc->data.u64BufAddr, cb, pDesc->data.cmd.fEOP /*fSend*/, fOnWorkerThread);
         }
         else
